requirements.txt
matplotlib==3.1.3
tqdm==4.42.1
pyyaml==5.4.1
torch==1.9.0
opencv-python==4.5.2.54
numpy==1.19.0
albumentations==1.0.0
torchvision==0.10.0
scikit-learn==0.24.2
pretrainedmodels==0.7.4
catalyst==21.5